Moving Dialog to its own module
