+++ tiledb_prompt_options/version.py
@@ -1,5 +1,5 @@
 # coding: utf-8
 # file generated by setuptools_scm
 # don't change, don't track in version control
-version = '0.1.dev5+dirty'
-version_tuple = (0, 1, 'dev5+dirty')
+version = '0.1.dev6+dirty'
+version_tuple = (0, 1, 'dev6+dirty')
